feat: polish code
